Customer and Game side both work fine now. Also, Login was referencing the wrong directories. Fixed.

+++ src/Customers/updateCustomer/deletecustomer.php
@@ -5,15 +5,22 @@ use App\Core\Database;
 include("../../../public/html/header.html");
 try{
     $db = Database::getInstance();
-    echo "Connection was Successful";
-    $sql = 'DELETE FROM customers WHERE cust_id = :cid';
-    $result = $db->prepare($sql);
-    $result->bindValue(':cid', $_POST['cust_id']);
-    $result->execute();
+    $cust_id = filter_input(INPUT_POST, 'cust_id', FILTER_VALIDATE_INT); // Gets the value from the URL.
+    if (!$cust_id) {
+        throw new Exception("Invalid Customer ID provided in the URL.");
+    }
+    $sql = 'DELETE FROM customers WHERE cust_id = :cid'; // Guess what this does.
+    $result = $db->query($sql, [':cid' => $cust_id]);
+    $rowCount = $result->rowCount();
+    if ($rowCount > 0) {
+        echo "Successfully deleted customer no: " . $_POST['cust_id'] . "<br>";
+    } else {
+        echo "No customer found with ID: " . $_POST['cust_id'] . "<br>";
+    }
     echo("You just deleted customer no: " . $_POST['cust_id'] ." \n click<a href='viewUpdateDelete.php'> here</a> to go back!");
 }catch(PDOException $e){
     if ($e->getCode() == 23000) {
-        echo "ooops couldnt delete as that record is linked to other tables click<a href='viewUpdateDelete.php'> here</a> to go back ";
+        echo "That Record has a dependency, and can't be deleted. Click<a href='viewUpdateDelete.php'> here</a> to go back!";
     }
 }
 ?>
